pyproject.toml
[tool.poetry]
name = "maze"
version = "0.1.0"
description = ""
authors = ["ryotaro <ryo.ex3991@gmail.com>"]
build = 'build.py'

[tool.poetry.dependencies]
python = "=3.9.6"
tabulate = "^0.8.9"
hexdump = "^3.3"
PyYAML = "^5.4.1"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
print-color = "^0.3.0"
pygame = "^2.0.1"
numpy = "^1.21.2"
Cython = "^0.29.24"
scipy = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[tool.poetry.scripts]
main = "main:run"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry>=0.12", "cython"]
build-backend = "poetry.core.masonry.api"